# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LabelHub",
    "description": "LabelHub MVP API foundation.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/tasks": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "List Tasks",
        "operationId": "list_tasks_tasks_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskRead"
                  },
                  "type": "array",
                  "title": "Response List Tasks Tasks Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Create Task",
        "operationId": "create_task_tasks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/tasks/{task_id}/template": {
      "get": {
        "tags": [
          "templates"
        ],
        "summary": "Get Template",
        "operationId": "get_template_tasks__task_id__template_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateSchemaRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}/template/draft": {
      "post": {
        "tags": [
          "templates"
        ],
        "summary": "Save Template Draft",
        "operationId": "save_template_draft_tasks__task_id__template_draft_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateDraftRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateSchemaRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}/template/publish": {
      "post": {
        "tags": [
          "templates"
        ],
        "summary": "Publish Template",
        "operationId": "publish_template_tasks__task_id__template_publish_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateSchemaRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckboxGroupField": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Id"
          },
          "label": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Label"
          },
          "required": {
            "type": "boolean",
            "title": "Required",
            "default": false
          },
          "helpText": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Helptext"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/TemplateOption"
            },
            "type": "array",
            "minItems": 1,
            "title": "Options"
          },
          "type": {
            "type": "string",
            "const": "checkbox_group",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "options",
          "type"
        ],
        "title": "CheckboxGroupField"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JsonField": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Id"
          },
          "label": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Label"
          },
          "required": {
            "type": "boolean",
            "title": "Required",
            "default": false
          },
          "helpText": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Helptext"
          },
          "type": {
            "type": "string",
            "const": "json",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "type"
        ],
        "title": "JsonField"
      },
      "LlmTool": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Id"
          },
          "label": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Label"
          },
          "promptTemplate": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 1,
            "title": "Prompttemplate"
          },
          "targetFieldId": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Targetfieldid"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "promptTemplate",
          "targetFieldId"
        ],
        "title": "LlmTool"
      },
      "LlmTriggerField": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Id"
          },
          "label": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Label"
          },
          "required": {
            "type": "boolean",
            "title": "Required",
            "default": false
          },
          "helpText": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Helptext"
          },
          "type": {
            "type": "string",
            "const": "llm_trigger",
            "title": "Type"
          },
          "promptTemplate": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 1,
            "title": "Prompttemplate"
          },
          "targetFieldId": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Targetfieldid"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "type",
          "promptTemplate",
          "targetFieldId"
        ],
        "title": "LlmTriggerField"
      },
      "NumberField": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Id"
          },
          "label": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Label"
          },
          "required": {
            "type": "boolean",
            "title": "Required",
            "default": false
          },
          "helpText": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Helptext"
          },
          "type": {
            "type": "string",
            "const": "number",
            "title": "Type"
          },
          "min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min"
          },
          "max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "type"
        ],
        "title": "NumberField"
      },
      "RadioField": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Id"
          },
          "label": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Label"
          },
          "required": {
            "type": "boolean",
            "title": "Required",
            "default": false
          },
          "helpText": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Helptext"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/TemplateOption"
            },
            "type": "array",
            "minItems": 1,
            "title": "Options"
          },
          "type": {
            "type": "string",
            "const": "radio",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "options",
          "type"
        ],
        "title": "RadioField"
      },
      "RatingField": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Id"
          },
          "label": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Label"
          },
          "required": {
            "type": "boolean",
            "title": "Required",
            "default": false
          },
          "helpText": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Helptext"
          },
          "type": {
            "type": "string",
            "const": "rating",
            "title": "Type"
          },
          "min": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Min",
            "default": 1
          },
          "max": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Max",
            "default": 5
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "type"
        ],
        "title": "RatingField"
      },
...
      "SelectField": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Id"
          },
          "label": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Label"
          },
          "required": {
            "type": "boolean",
            "title": "Required",
            "default": false
          },
          "helpText": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Helptext"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/TemplateOption"
            },
            "type": "array",
            "minItems": 1,
            "title": "Options"
          },
          "type": {
            "type": "string",
            "const": "select",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "options",
          "type"
        ],
        "title": "SelectField"
      },
      "ShowItemField": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Id"
          },
          "label": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Label"
          },
          "required": {
            "type": "boolean",
            "title": "Required",
            "default": false
          },
          "helpText": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Helptext"
          },
          "type": {
            "type": "string",
            "const": "show_item",
            "title": "Type"
          },
          "source": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "type",
          "source"
        ],
        "title": "ShowItemField"
      },
...
      "TaskCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "distribution_strategy": {
            "type": "string",
            "title": "Distribution Strategy",
            "default": "manual"
          },
          "quota_per_labeler": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Quota Per Labeler"
          },
          "deadline_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline At"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "TaskCreate"
      },
...
      "TaskRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "distribution_strategy": {
            "type": "string",
            "title": "Distribution Strategy"
          },
          "quota_per_labeler": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quota Per Labeler"
          },
          "deadline_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline At"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "status",
          "distribution_strategy",
          "quota_per_labeler",
          "deadline_at",
          "created_by",
          "created_at",
          "updated_at"
        ],
        "title": "TaskRead"
      },
      "TaskStatus": {
        "type": "string",
        "enum": [
          "draft",
          "published",
          "paused",
          "ended"
        ],
        "title": "TaskStatus"
      },
...
      "TemplateDocument": {
        "properties": {
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version",
            "default": 1
          },
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "layout": {
            "$ref": "#/components/schemas/TemplateLayout"
          },
          "fields": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ShowItemField"
                },
                {
                  "$ref": "#/components/schemas/TextField"
                },
                {
                  "$ref": "#/components/schemas/TextareaField"
                },
                {
                  "$ref": "#/components/schemas/NumberField"
                },
                {
                  "$ref": "#/components/schemas/RadioField"
                },
                {
                  "$ref": "#/components/schemas/CheckboxGroupField"
                },
                {
                  "$ref": "#/components/schemas/SelectField"
                },
                {
                  "$ref": "#/components/schemas/RatingField"
                },
                {
                  "$ref": "#/components/schemas/JsonField"
                },
                {
                  "$ref": "#/components/schemas/LlmTriggerField"
                }
              ],
              "discriminator": {
                "propertyName": "type",
                "mapping": {
                  "checkbox_group": "#/components/schemas/CheckboxGroupField",
                  "json": "#/components/schemas/JsonField",
                  "llm_trigger": "#/components/schemas/LlmTriggerField",
                  "number": "#/components/schemas/NumberField",
                  "radio": "#/components/schemas/RadioField",
                  "rating": "#/components/schemas/RatingField",
                  "select": "#/components/schemas/SelectField",
                  "show_item": "#/components/schemas/ShowItemField",
                  "text": "#/components/schemas/TextField",
                  "textarea": "#/components/schemas/TextareaField"
                }
              }
            },
            "type": "array",
            "minItems": 1,
            "title": "Fields"
          },
          "llmTools": {
            "items": {
              "$ref": "#/components/schemas/LlmTool"
            },
            "type": "array",
            "title": "Llmtools"
          },
          "validations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Validations"
          },
          "visibilityRules": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Visibilityrules"
          }
        },
        "type": "object",
        "required": [
          "title",
          "fields"
        ],
        "title": "TemplateDocument"
      },
      "TemplateDraftRequest": {
        "properties": {
          "schema": {
            "$ref": "#/components/schemas/TemplateDocument"
          }
        },
        "type": "object",
        "required": [
          "schema"
        ],
        "title": "TemplateDraftRequest"
      },
      "TemplateLayout": {
        "properties": {
          "type": {
            "type": "string",
            "const": "single",
            "title": "Type",
            "default": "single"
          },
          "groups": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Groups"
          }
        },
        "type": "object",
        "title": "TemplateLayout"
      },
      "TemplateOption": {
        "properties": {
          "label": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Label"
          },
          "value": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "label",
          "value"
        ],
        "title": "TemplateOption"
      },
      "TemplateSchemaRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "schema_payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Schema Payload"
          },
          "is_published": {
            "type": "boolean",
            "title": "Is Published"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "task_id",
          "version",
          "title",
          "schema_payload",
          "is_published",
          "created_by",
          "created_at",
          "published_at"
        ],
        "title": "TemplateSchemaRead"
      },
      "TextField": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Id"
          },
          "label": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Label"
          },
          "required": {
            "type": "boolean",
            "title": "Required",
            "default": false
          },
          "helpText": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Helptext"
          },
          "type": {
            "type": "string",
            "const": "text",
            "title": "Type"
          },
          "placeholder": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Placeholder"
          },
          "minLength": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Minlength"
          },
          "maxLength": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxlength"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "type"
        ],
        "title": "TextField"
      },
      "TextareaField": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Id"
          },
          "label": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Label"
          },
          "required": {
            "type": "boolean",
            "title": "Required",
            "default": false
          },
          "helpText": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Helptext"
          },
          "type": {
            "type": "string",
            "const": "textarea",
            "title": "Type"
          },
          "placeholder": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Placeholder"
          },
          "minLength": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Minlength"
          },
          "maxLength": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxlength"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "type"
        ],
        "title": "TextareaField"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}